requirements.txt
alembic==1.13.2
asyncpg==0.29.0
bcrypt==4.2.0
fastapi==0.112.0
greenlet==3.0.3
gunicorn==22.0.0
loguru==0.7.2
psutil==6.0.0
pydantic==2.8.2
PyJWT==2.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
SQLAlchemy==2.0.31
starlette>=0.37.2,<0.38.0
typing_extensions==4.12.2
uvicorn==0.30.5
cryptography==43.0.0
python-multipart==0.0.9